Update research card generation
