+++ research.html
@@ -171,12 +171,12 @@
           <div class="row">
             <div class="col-lg-12">
               <div class="course-organization-name d-flex flex-column align-items-start h-100 justify-content-start">
-                <p class="section-title lead display-6 fw-bold">Conferences</p>
+                <p class="section-title lead display-6 fw-bold">Applied Mathematics</p>
               </div>
             </div>
           </div>
 
-          <!-- Oral Subtopic with Search Field -->
+          <!-- Oral Subtopic-->
           <div class="row pt-md-5 pb-md-5">
             <div class="col-lg-12 d-flex align-items-center">
               <div
@@ -195,7 +195,7 @@
             </div>
           </div>
 
-          <!-- Oral Subtopic  -->
+          <!-- Poster Subtopic  -->
           <div class="row pt-md-5 pb-md-5">
             <div class="col-lg-12 d-flex align-items-center">
               <div
@@ -391,7 +391,7 @@
     crossorigin="anonymous"></script>
 
   <script src="javascript/navbar.js"></script>
-  <script src="javascript/research.js"></script>
+  <script type="module" src="javascript/research/research.js"></script>
 </body>
 
 </html>
